logstash setup modified
Changed the way logstash is setup so it binds all its ports to the gateway ip address that way all containers can just log to 172.17.0.1:ports with out requiring a dependency on the logstash container. Makes terraform plans better since if we had it like it was before it would force terra to destroy all containers if logstash had to be changed. Not good.